Create README.md
Mobile-Facebook-Drive
=====================

Mobile Facebook Drive is a store for pictures, comments of facebook on WP7. This is similar application to SkyDrive 
applicaton. Idea is to create an application which stores all the cloud services pictures such as Picassa, Facebook,
Flickr, SkyDrive and other for Mobile devices.

This initial version integrated with Facebook and a part of Flickr. 

The following features it is been supported.

– Silverlight app for WP7 (Mango) with modern(Metro) UI 
– Model-View-ViewModel architecture 
– login to facebook                          
– logout to facebook                        
– take a picture with cameta                
– upload pictures  to facebook               
– download pictures  from facebook             
– review one single picture       
– review the collection (a set of pictures) 
– WP7 navigation                     
– tombstoning handling               
– live tile                           
– application bar                    
– splash screen support              
– orientations support 
– tappable support							
- localization

- A part of flicker login and authentication has done.